Finalizados os testes do pacote remote_ref

+++ include/remote_ref.h
@@ -18,4 +18,4 @@ typedef struct remote_ref{
 
 remote_ref *init_remote_ref(const char *addr_str);
 void free_remote_ref(remote_ref *r);
-int confrontation(remote_ref *remote_a, remote_ref *remote_b);
+uint8_t confrontation(remote_ref *remote_a, remote_ref *remote_b);
